+++ examples/compute/create_instance_with_volumes/main.go
@@ -1,6 +1,7 @@
 //
 // Copyright 2020 Joyent, Inc. All rights reserved.
 // Copyright 2025 MNX Cloud, Inc.
+// Copyright 2026 Edgecast Cloud LLC.
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -180,6 +181,8 @@ func main() {
 
 	// Create a new instance using our input attributes...
 	// https://github.com/TritonDataCenter/triton-go/v2/blob/master/compute/instances.go#L206
+	tagsInput := make(map[string]interface{}, 1)
+	tagsInput["tag1"] = "value1"
 	createInput := &compute.CreateInstanceInput{
 		Name:     testutils.RandString(10),
 		Package:  PackageName,
@@ -188,9 +191,7 @@ func main() {
 		Volumes: []compute.InstanceVolume{
 			instanceVolume,
 		},
-		Tags: map[string]string{
-			"tag1": "value1",
-		},
+		Tags: tagsInput,
 	}
 
 	created, err := c.Instances().Create(context.Background(), createInput)