+++ examples/storage/force_put/main.go
@@ -1,6 +1,7 @@
 //
 // Copyright 2020 Joyent, Inc. All rights reserved.
 // Copyright 2025 MNX Cloud, Inc.
+// Copyright 2026 Edgecast Cloud LLC.
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -25,10 +26,10 @@ import (
 )
 
 func main() {
-	keyID := os.Getenv("TRITON_KEY_ID")
-	keyMaterial := os.Getenv("TRITON_KEY_MATERIAL")
+	keyID := os.Getenv("MANTA_KEY_ID")
+	keyMaterial := os.Getenv("MANTA_KEY_MATERIAL")
 	mantaUser := os.Getenv("MANTA_USER")
-	userName := os.Getenv("TRITON_USER")
+	userName := os.Getenv("MANTA_SUBUSER")
 
 	var signer authentication.Signer
 	var err error