+++ examples/storage/list_directory/main.go
@@ -1,6 +1,7 @@
 //
 // Copyright 2020 Joyent, Inc. All rights reserved.
 // Copyright 2025 MNX Cloud, Inc.
+// Copyright 2026 Edgecast Cloud LLC.
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -32,7 +33,7 @@ func main() {
 		keyID       = os.Getenv("MANTA_KEY_ID")
 		accountName = os.Getenv("MANTA_USER")
 		keyMaterial = os.Getenv("MANTA_KEY_MATERIAL")
-		userName    = os.Getenv("TRITON_USER")
+		userName    = os.Getenv("MANTA_SUBUSER")
 	)
 
 	if keyMaterial == "" {