+++ storage/directory.go
@@ -1,6 +1,7 @@
 //
 // Copyright 2020 Joyent, Inc. All rights reserved.
 // Copyright 2025 MNX Cloud, Inc.
+// Copyright 2026 Edgecast Cloud LLC.
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -132,27 +133,22 @@ func (s *DirectoryClient) Put(ctx context.Context, input *PutDirectoryInput) err
 // DeleteDirectoryInput represents parameters to a Delete operation.
 type DeleteDirectoryInput struct {
 	DirectoryName string
-	ForceDelete   bool //Will recursively delete all child directories and objects
+	ForceDelete   bool // Recursively delete all child directories and objects
 }
 
-// Delete deletes a directory on the Triton Object Storage. The directory must
-// be empty.
+// Delete deletes a directory on the Triton Object Storage. If ForceDelete is
+// false, the directory must be empty.
 func (s *DirectoryClient) Delete(ctx context.Context, input *DeleteDirectoryInput) error {
 	absPath := absFileInput(s.client.AccountName, input.DirectoryName)
 
 	if input.ForceDelete {
-		err := deleteAll(*s, ctx, absPath)
-		if err != nil {
+		if err := deleteAll(*s, ctx, absPath); err != nil {
 			return err
 		}
-	} else {
-		err := deleteDirectory(*s, ctx, absPath)
-		if err != nil {
-			return err
-		}
+		return deleteDirectory(*s, ctx, absPath)
 	}
 
-	return nil
+	return deleteDirectory(*s, ctx, absPath)
 }
 
 func deleteAll(c DirectoryClient, ctx context.Context, directoryPath _AbsCleanPath) error {
@@ -165,12 +161,16 @@ func deleteAll(c DirectoryClient, ctx context.Context, directoryPath _AbsCleanPa
 	for _, obj := range objs.Entries {
 		newPath := absFileInput(c.client.AccountName, path.Join(string(directoryPath), obj.Name))
 		if obj.Type == "directory" {
-			err := deleteDirectory(c, ctx, newPath)
-			if err != nil {
-				return deleteAll(c, ctx, newPath)
+			if err := deleteAll(c, ctx, newPath); err != nil {
+				return err
+			}
+			if err := deleteDirectory(c, ctx, newPath); err != nil {
+				return err
 			}
 		} else {
-			return deleteObject(c, ctx, newPath)
+			if err := deleteObject(c, ctx, newPath); err != nil {
+				return err
+			}
 		}
 	}