+++ testutils/steps.go
@@ -1,6 +1,7 @@
 //
 // Copyright 2020 Joyent, Inc.
 // Copyright 2025 MNX Cloud, Inc.
+// Copyright 2026 Edgecast Cloud LLC.
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -138,7 +139,7 @@ type StepGetImage struct {
 
 func (s *StepGetImage) Run(state TritonStateBag) StepAction {
 	const imageName = "ubuntu-24.04"
-	const imageVersion = "20240612"
+	const imageVersion = "20260121"
 
 	computeClient, err := compute.NewClient(state.Config())
 	if err != nil {
@@ -156,8 +157,8 @@ func (s *StepGetImage) Run(state TritonStateBag) StepAction {
 	}
 
 	if len(images) == 0 {
-		state.AppendError(pkgerrors.Errorf("No images matching image name %s",
-			imageName))
+		state.AppendError(pkgerrors.Errorf("No images matching image name %s v%s",
+			imageName, imageVersion))
 		return Halt
 	}